refactor: 테스트를 DB에 의존하지 않고 동작하도록 수정
