+++ src/main/java/chess/dao/TurnDao.java
@@ -1,41 +1,12 @@
 package chess.dao;
 
-import chess.db.JdbcTemplate;
-import chess.db.RowMapper;
 import chess.dto.TurnDto;
 
-import java.util.List;
+public interface TurnDao {
 
-public class TurnDao {
-    private static final String TABLE_NAME = "turns";
+    TurnDto findOne();
 
-    private final JdbcTemplate jdbcTemplate;
-    private final RowMapper<TurnDto> rowMapper = (resultSet) ->
-            new TurnDto(resultSet.getString("turn"));
+    void update(final TurnDto turnDto);
 
-    public TurnDao(final JdbcTemplate jdbcTemplate) {
-        this.jdbcTemplate = jdbcTemplate;
-    }
-
-    public TurnDto findOne() {
-        final String query = "SELECT * FROM " + TABLE_NAME + " LIMIT 1";
-        final List<TurnDto> turns = jdbcTemplate.executeAndGet(query, rowMapper);
-        if (turns.isEmpty()) {
-            throw new IllegalArgumentException("데이터가 없습니다.");
-        }
-        return turns.get(0);
-    }
-
-
-    public void update(final TurnDto turnDto) {
-        final String deleteQuery = "DELETE FROM " + TABLE_NAME;
-        final String insertQuery = "INSERT INTO " + TABLE_NAME + " VALUES (?)";
-        jdbcTemplate.execute(deleteQuery);
-        jdbcTemplate.execute(insertQuery, turnDto.turn());
-    }
-
-    public void deleteAll() {
-        final String query = "DELETE FROM " + TABLE_NAME;
-        jdbcTemplate.execute(query);
-    }
+    void deleteAll();
 }
