+++ src/test/java/chess/dao/PieceDaoTest.java
@@ -1,6 +1,5 @@
 package chess.dao;
 
-import chess.db.JdbcTemplate;
 import chess.dto.PieceDto;
 import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.BeforeEach;
@@ -10,7 +9,7 @@ import org.junit.jupiter.api.Test;
 import static org.assertj.core.api.Assertions.assertThat;
 
 class PieceDaoTest {
-    private PieceDao pieceDao = new PieceDao(new JdbcTemplate());
+    private PieceDao pieceDao = new PieceSimpleDao();
 
     @BeforeEach
     void setUp() {
