+++ src/test/java/chess/dao/TurnDaoTest.java
@@ -1,6 +1,5 @@
 package chess.dao;
 
-import chess.db.JdbcTemplate;
 import chess.domain.piece.Team;
 import chess.dto.TurnDto;
 import org.junit.jupiter.api.DisplayName;
@@ -13,8 +12,8 @@ class TurnDaoTest {
     @Test
     @DisplayName("턴을 저장할 수 있다.")
     void update() {
+        TurnDao turnDao = new TurnSimpleDao();
         TurnDto whiteTurn = TurnDto.from(Team.WHITE);
-        TurnDao turnDao = new TurnDao(new JdbcTemplate());
 
         turnDao.update(whiteTurn);
